+++ code/CompilerLab/analyzer.h
@@ -12,7 +12,7 @@ private:
 	FuncSymbolTable func_symbol_table;
 private:
 	uint AllocateVarIndex(uint length) { uint index = var_index_stack.back(); var_index_stack.back() += length; return index; }
-	uint AllocateFuncIndex() { return func_symbol_table.size(); }
+	uint AllocateFuncIndex() { return (uint)func_symbol_table.size(); }
 private:
 	const VarEntry& AddVar(string_view identifier, const ArraySize& array_size, bool is_global, bool is_parameter);
 	void AddConstVar(string_view identifier, const ArraySize& array_size, const InitializingList& initializing_list);
